+++ libs/@guardian/libs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @guardian/libs
 
+## 11.1.0
+
+### Minor Changes
+
+- 5a7ba38: Add `isNonNullable` [type guard](https://www.typescriptlang.org/docs/handbook/advanced-types.html#type-guards-and-differentiating-types) for values that are neither `null` nor `undefined`
+
+## 11.0.0
+
+### Major Changes
+
+- eb65fde: Add `typescript@^4.3.2` to `peerDependecies` (`optional`).
+
+  This should mean you get a warning if you're using an incompatible version of typescript, but not if you aren't using it at all.
+
 ## 10.1.1
 
 ### Patch Changes